requirements.txt
google-generativeai
SpeechRecognition
pynput
pyttsx3
python-dotenv 